Core: Movie uniqueId into its own module
