+++ ui/accordion.js
@@ -23,8 +23,9 @@
 		define( [
 			"jquery",
 			"./version",
-			"./core",
+			"./version",
 			"./keycode",
+			"./unique-id",
 			"./widget"
 		], factory );
 	} else {
